ShowTickets activity buytickets runnable class added. Still some problems, namely, could not load correctly the spinners from the new_ticket xml. Some details updated in Passenger, RestApi and LoginActivity classes.

+++ passenger_app/src/Entities/Passenger.java
@@ -49,7 +49,7 @@ public class Passenger {
 		return creditCardNumber;
 	}
 
-	public void setCreditCardNumber(int creditCardNumber) {
+	public void setCreditCardNumber(long creditCardNumber) {
 		this.creditCardNumber = creditCardNumber;
 	}
